+++ sdk/cosmos/azure-cosmos/src/main/java/com/azure/cosmos/CosmosClientBuilder.java
@@ -1051,7 +1051,6 @@ public class CosmosClientBuilder implements
         buildConnectionPolicy();
         CosmosAsyncClient cosmosAsyncClient = new CosmosAsyncClient(this);
         if (proactiveContainerInitConfig != null) {
-
             cosmosAsyncClient.recordOpenConnectionsAndInitCachesStarted(proactiveContainerInitConfig.getCosmosContainerIdentities());
 
             Duration aggressiveWarmupDuration = proactiveContainerInitConfig
@@ -1191,7 +1190,7 @@ public class CosmosClientBuilder implements
 
             DiagnosticsProvider provider = client.getDiagnosticsProvider();
             if (provider != null) {
-                tracingCfg = provider.isEnabled() + ", " + provider.isRealTracer();
+                tracingCfg = provider.getTraceConfigLog();
             }
 
             // NOTE: if changing the logging below - do not log any confidential info like master key credentials etc.